change security structure and add more changes

+++ src/main/java/com/sebi/controller/CartItemController.java
@@ -1,19 +1,21 @@
 package com.sebi.controller;
 
+import com.sebi.config.JwtService;
 import com.sebi.exception.CartItemException;
-import com.sebi.exception.OrderException;
 import com.sebi.exception.UserException;
 import com.sebi.model.CartItem;
 import com.sebi.model.User;
 import com.sebi.repository.CartItemRepository;
+import com.sebi.repository.UserRepository;
 import com.sebi.response.ApiResponse;
 import com.sebi.service.CartItemService;
-import com.sebi.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/cart_items")
@@ -21,7 +23,9 @@ public class CartItemController {
     @Autowired
     private CartItemService cartItemService;
     @Autowired
-    private UserService userService;
+    private UserRepository userRepository;
+    @Autowired
+    private JwtService jwtService;
     @Autowired
     private CartItemRepository cartItemRepository;
 
@@ -30,23 +34,36 @@ public class CartItemController {
     @Operation(description = "Remove Cart Item From Cart")
     public ResponseEntity<ApiResponse> DeleteOrderHandler(@PathVariable Long itemId,
                                                           @RequestHeader("Authorization") String jwt) throws UserException, CartItemException {
-        User user = userService.findUserProfileByJwt(jwt);
-        Long userId = user.getId();
-        cartItemService.removeCartItem(userId,itemId);
-        ApiResponse res = new ApiResponse();
-        res.setMessage("item deleted successfully");
-        res.setStatus(true);
+        String token = jwtService.extractBearer(jwt);
+        String username = jwtService.extractUserName(token);
+        Optional<User> user = userRepository.findByUsername(username);
+        if(user.isPresent()){
+            User newUser = user.get();
+            Long userId = newUser.getId();
+            cartItemService.removeCartItem(userId,itemId);
+            ApiResponse res = new ApiResponse();
+            res.setMessage("item deleted successfully");
+            res.setStatus(true);
 
-        return new ResponseEntity<>(res, HttpStatus.OK);
+            return new ResponseEntity<>(res, HttpStatus.OK);
+        }else {
+            throw new UserException("User not found!");
+        }
     }
 
     @PutMapping("/{cartItemId}")
     @Operation(description = "Update Item To Cart")
     public ResponseEntity<CartItem> updateCartItem(@RequestBody CartItem cartItem,
                                                    @PathVariable Long cartItemId, @RequestHeader("Authorization") String jwt) throws UserException,CartItemException{
-        User user = userService.findUserProfileByJwt(jwt);
-        CartItem updatedCartItem = cartItemService.updateCartItem(user.getId(),cartItemId,cartItem);
-
-        return new ResponseEntity<>(updatedCartItem,HttpStatus.OK);
+        String token = jwtService.extractBearer(jwt);
+        String username = jwtService.extractUserName(token);
+        Optional<User> user = userRepository.findByUsername(username);
+        if(user.isPresent()){
+            User newUser = user.get();
+            CartItem updatedCartItem = cartItemService.updateCartItem(newUser.getId(),cartItemId,cartItem);
+            return new ResponseEntity<>(updatedCartItem,HttpStatus.OK);
+        }else {
+            throw new UserException("User not found!");
+        }
     }
 }
